requirements.txt
streamlit
pandas
matplotlib
seaborn
scikit-learn
nltk
vaderSentiment
Pillow
